requirements.txt
# Core dependencies
setuptools>=68.0.0  # Required for pkg_resources
streamlit==1.31.1
alpaca-py==0.10.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
requests==2.31.0
plotly==5.18.0

# Async/websocket support
aiohttp>=3.8.4  # For async HTTP requests
websockets>=11.0.3  # Required for async websocket support
websocket-client==1.6.4  # For synchronous websocket fallback

# Technical analysis
ta==0.11.0
pandas_ta==0.3.14b0  # Pinned version for numpy compatibility
TA-Lib==0.6.3  # Updated to match installed version

# API integrations
finnhub-python>=2.4.20
openai>=1.3.0

# Data processing and ML
scikit-learn==1.3.2
statsmodels>=0.14.0
lightgbm>=4.0.0
xgboost>=2.0.0
prophet>=1.1.4
optuna>=3.3.0

# Optional ML components (uncomment if needed)
# tensorflow>=2.13.0  # Requires Python <=3.11

# UI enhancements
streamlit-extras>=0.3.0
streamlit-autorefresh==1.0.0
streamlit-option-menu>=0.3.2
streamlit-aggrid>=0.3.4

# Testing
pytest==7.4.3
pytest-mock>=3.11.0

# Utilities
colorlog>=6.7.0
pytz==2023.3.post1
ccxt>=3.1.0   # For crypto exchange data
joblib>=1.3.2 # For model persistence
tqdm>=4.66.1  # For progress bars

# Added from the code block
matplotlib==3.8.2
pymongo==4.6.1
pydantic==2.5.2
scipy==1.11.4

# New dependencies
python-binance==1.0.19
ta-lib==0.4.28 